[2127] Project restructured for feature based package
