+++ slack-compose-sample/src/main/java/io/getstream/compose/slack/ThemeConfig.kt
@@ -6,6 +6,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
 import io.getstream.chat.android.compose.ui.theme.StreamShapes
 
+/**
+ * This class can be added with more Theme customizations for ChatTheme, such as [StreamShapes]
+ */
+
 @Composable
 fun shapes(): StreamShapes {
     return StreamShapes(
