+++ slack-compose-sample/src/main/java/io/getstream/compose/slack/ui/features/profile/ProfileScreen.kt
@@ -1,23 +1,17 @@
-package io.getstream.compose.slack.presentation.screens
+package io.getstream.compose.slack.ui.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
-import io.getstream.compose.slack.presentation.components.FullAvatar
-import io.getstream.compose.slack.presentation.components.StatusInput
+import io.getstream.compose.slack.ui.features.profile.FullAvatar
+import io.getstream.compose.slack.ui.components.StatusInput
 
 /** [WIP]
  * A screen component to represent profile or a settings page for the slack workspace example.
